hospitalmanagement/requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.6.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cookiecutter==2.7.1
cryptography==46.0.5
Django==6.0.3
django-tailwind-4==0.1.4
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
